fix conflict with offer credential branch
