+++ src/helpers/mod.rs
@@ -1,11 +1,8 @@
 mod version_info;
+
 #[cfg(feature = "plugin-vc-zkp-bbs")]
 mod credential;
-#[cfg(feature = "plugin-vc-zkp-bbs")]
-mod credential_merge_with_pr33_later_on;
 
 #[cfg(feature = "plugin-vc-zkp-bbs")]
 pub(crate) use credential::Credential;
-#[cfg(feature = "plugin-vc-zkp-bbs")]
-pub(crate) use credential_merge_with_pr33_later_on::Credential;
 pub(crate) use version_info::VersionInfo;
